feat: move /bin/template/cordova/lib to /lib
