+++ lib/clean.js
